Split out useOIDC into a more generic form, and make useZemnMeAuth a user of it.
