+++ project/zemn.me/app/admin/client.tsx
@@ -8,11 +8,11 @@ import { z } from 'zod';
 import type { components } from '#root/project/zemn.me/api/api_client.gen.js';
 import Link from '#root/project/zemn.me/components/Link/index.js';
 import { PendingPip } from '#root/project/zemn.me/components/PendingPip/PendingPip.js';
-import { useOIDC } from '#root/project/zemn.me/hook/useOIDC.js';
 import {
 	useGetAdminUid,
 	useZemnMeApi,
 } from '#root/project/zemn.me/hook/useZemnMeApi.js';
+import { useZemnMeAuth } from '#root/project/zemn.me/hook/useZemnMeAuth.js';
 import {
 	and_then as option_and_then,
 	is_some as option_is_some,
@@ -447,7 +447,7 @@ function DisplayAdminUid({
 }
 
 export default function Admin() {
-	const [idToken, promptForLogin] = useOIDC();
+	const [idToken, promptForLogin] = useZemnMeAuth();
 	const loginReady = option_is_some(promptForLogin);
 
 	const handleLogin = () => {
