+++ project/zemn.me/hook/BUILD.bazel
@@ -13,6 +13,7 @@ ts_project(
         "useOpenIDConfiguration.ts",
         "useWindowCallback.ts",
         "useZemnMeApi.tsx",
+        "useZemnMeAuth.tsx",
     ],
     visibility = [
         "//project/zemn.me:__subpackages__",
@@ -55,6 +56,7 @@ filegroup(
         "useOpenIDConfiguration.ts",
         "useWindowCallback.ts",
         "useZemnMeApi.tsx",
+        "useZemnMeAuth.tsx",
     ],
     visibility = ["//:__subpackages__"],
 )
